Thanks everyone for helping getting this together, especially Alessio
and DLite :-)

+++ flight/libraries/inc/CoordinateConversions.h
@@ -31,12 +31,12 @@
 #define COORDINATECONVERSIONS_H_
 
 // ****** convert Lat,Lon,Alt to ECEF  ************
-void LLA2ECEF(float LLA[3], float ECEF[3]);
+void LLA2ECEF(float LLA[3], double ECEF[3]);
 
 // ****** convert ECEF to Lat,Lon,Alt (ITERATIVE!) *********
-uint16_t ECEF2LLA(float ECEF[3], float LLA[3]);
+uint16_t ECEF2LLA(double ECEF[3], float LLA[3]);
 
-void RneFromLLA(float LLA[3], float Rne[3][3]);
+void RneFromLLA(float LLA[3], double Rne[3][3]);
 
 // ****** find rotation matrix from rotation vector
 void Rv2Rot(float Rv[3], float R[3][3]);
@@ -51,10 +51,10 @@ void RPY2Quaternion(const float rpy[3], float q[4]);
 void Quaternion2R(float q[4], float Rbe[3][3]);
 
 // ****** Express LLA in a local NED Base Frame ********
-void LLA2Base(float LLA[3], float BaseECEF[3], float Rne[3][3], float NED[3]);
+void LLA2Base(float LLA[3], double BaseECEF[3], double Rne[3][3], float NED[3]);
 
 // ****** Express ECEF in a local NED Base Frame ********
-void ECEF2Base(float ECEF[3], float BaseECEF[3], float Rne[3][3], float NED[3]);
+void ECEF2Base(double ECEF[3], double BaseECEF[3], double Rne[3][3], float NED[3]);
 
 // ****** convert Rotation Matrix to Quaternion ********
 // ****** if R converts from e to b, q is rotation from e to b ****
